ci: make black, isort, pydocstyle non-local hooks
These packages do not require additional dependencies as defined in the
setup.cfg. Originally, these were local hooks to ensure that these
hooks behave the same as the versions of black/isort/pydocstyle that
were installed as optional requirements (through pip install -e .[dev]).

Note: there is no guarantee that the package versions of these hooks
are the same as the pinned versions from the constraints files. The
upgrade pinning job should ensure that (only problem being that
packages in setup.cfg can be constraint to lower versions, while
the pre-commit hooks always upgrade to the latest version).